test: update tests to reflect move away from Tokens class
